requirements.txt
flask==3.0.0
requests==2.31.0
pydantic==2.5.2
typing-extensions==4.9.0
python-dotenv==1.0.0
jsonschema==4.20.0
ollama==0.1.7